openmp req

+++ src/ngraph/runtime/cpu/kernel/gather.hpp
@@ -154,7 +154,7 @@ namespace ngraph
 #ifdef _OPENMP
 #pragma omp parallel for
 #endif
-                        for (size_t i = 0; i < outer_loop_num * num_indices; i++)
+                        for (int64_t i = 0; i < outer_loop_num * num_indices; i++)
                         {
                             Eigen::array<Eigen::Index, Rank1> in_extents, in_offsets;
                             Eigen::array<Eigen::Index, Rank2> out_extents, out_offsets;
